+++ src/type/IterableType.php
@@ -70,4 +70,9 @@ final class IterableType extends Type
     {
         return $this->allowsNull;
     }
+
+    public function isIterable(): bool
+    {
+        return true;
+    }
 }