+++ src/type/StaticType.php
@@ -51,4 +51,9 @@ final class StaticType extends Type
     {
         return $this->allowsNull;
     }
+
+    public function isStatic(): bool
+    {
+        return true;
+    }
 }