+++ tests/unit/type/IterableTypeTest.php
@@ -82,4 +82,19 @@ final class IterableTypeTest extends TestCase
             )
         );
     }
+
+    public function testCanBeQueriedForType(): void
+    {
+        $this->assertFalse($this->type->isCallable());
+        $this->assertFalse($this->type->isGenericObject());
+        $this->assertTrue($this->type->isIterable());
+        $this->assertFalse($this->type->isMixed());
+        $this->assertFalse($this->type->isNull());
+        $this->assertFalse($this->type->isObject());
+        $this->assertFalse($this->type->isSimple());
+        $this->assertFalse($this->type->isStatic());
+        $this->assertFalse($this->type->isUnion());
+        $this->assertFalse($this->type->isUnknown());
+        $this->assertFalse($this->type->isVoid());
+    }
 }