+++ tests/unit/type/SimpleTypeTest.php
@@ -138,4 +138,21 @@ final class SimpleTypeTest extends TestCase
     {
         $this->assertSame('string', Type::fromValue('string', false)->value());
     }
+
+    public function testCanBeQueriedForType(): void
+    {
+        $type = new SimpleType('bool', false);
+
+        $this->assertFalse($type->isCallable());
+        $this->assertFalse($type->isGenericObject());
+        $this->assertFalse($type->isIterable());
+        $this->assertFalse($type->isMixed());
+        $this->assertFalse($type->isNull());
+        $this->assertFalse($type->isObject());
+        $this->assertTrue($type->isSimple());
+        $this->assertFalse($type->isStatic());
+        $this->assertFalse($type->isUnion());
+        $this->assertFalse($type->isUnknown());
+        $this->assertFalse($type->isVoid());
+    }
 }