+++ tests/unit/type/UnknownTypeTest.php
@@ -13,6 +13,8 @@ use PHPUnit\Framework\TestCase;
 
 /**
  * @covers \SebastianBergmann\Type\UnknownType
+ *
+ * @uses \SebastianBergmann\Type\Type
  */
 final class UnknownTypeTest extends TestCase
 {
@@ -59,5 +61,20 @@ final class UnknownTypeTest extends TestCase
     public function testCanBeRepresentedAsString(): void
     {
         $this->assertSame('', $this->type->asString());
+    }
+
+    public function testCanBeQueriedForType(): void
+    {
+        $this->assertFalse($this->type->isCallable());
+        $this->assertFalse($this->type->isGenericObject());
+        $this->assertFalse($this->type->isIterable());
+        $this->assertFalse($this->type->isMixed());
+        $this->assertFalse($this->type->isNull());
+        $this->assertFalse($this->type->isObject());
+        $this->assertFalse($this->type->isSimple());
+        $this->assertFalse($this->type->isStatic());
+        $this->assertFalse($this->type->isUnion());
+        $this->assertTrue($this->type->isUnknown());
+        $this->assertFalse($this->type->isVoid());
     }
 }